+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.11.8] - 2020-02-26
+### Fixed
+- 中文域名 Mapping Error Bug
+- 優化 Check-sync All DNS Pod Record 效能
+
 ## [0.11.7] - 2020-02-14
 ### Fixed
 - Delete Domain Error Message